change constant of ele, try to remove intake and collecter
