+++ src/main/java/frc/robot/commands/EjectIntakeCommand.java
@@ -5,36 +5,31 @@
 package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.Collector.Collector;
+import frc.robot.subsystems.Collector.Feeder.Feeder;
 
 /* You should consider using the more terse Command factories API instead https://docs.wpilib.org/en/stable/docs/software/commandbased/organizing-command-based.html#defining-commands */
 public class EjectIntakeCommand extends Command {
-  private Collector m_Collector;
+  private Feeder m_feeder;
   /** Creates a new EjectIntakeCommand. */
-  public EjectIntakeCommand(Collector collector) {
-    m_Collector = collector;
+  public EjectIntakeCommand(Feeder feeder) {
+    m_feeder = feeder;
     // Use addRequirements() here to declare subsystem dependencies.
   }
 
   // Called when the command is initially scheduled.
   @Override
-  public void initialize() {
-    m_Collector.extendIntake();
-  }
+  public void initialize() {}
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    m_Collector.runIntake(-12.0);
-    m_Collector.runFeeder(-12.0);
+    m_feeder.runFeeder(-12.0);
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    m_Collector.stopIntake();
-    m_Collector.stopFeeder();
-    m_Collector.retractIntake();
+    m_feeder.stop();
   }
 
   // Returns true when the command should end.
